# Optimized in-memory datagrid mount performance