--- a/src/components/datagrid/data_grid.tsx
+++ b/src/components/datagrid/data_grid.tsx
@@ -379,28 +379,46 @@ function useInMemoryValues(
   rowCount: number
 ): [
   EuiDataGridInMemoryValues,
-  (rowIndex: number, column: EuiDataGridColumn, value: string) => void
+  (rowIndex: number, columnId: string, value: string) => void
 ] {
-  const [inMemoryValues, setInMemoryValues] = useState<
-    EuiDataGridInMemoryValues
-  >({});
+  /**
+   * For performance, `onCellRender` below mutates the inMemoryValues object
+   * instead of cloning. If this operation were done in a setState call
+   * React would ignore the update as the object itself has not changed.
+   * So, we keep a dual record: the in-memory values themselves and a "version" counter.
+   * When the object is mutated, the version is incremented triggering a re-render, and
+   * the returned `inMemoryValues` object is re-created (cloned) from the mutated version.
+   * The version updates are batched, so only one clone happens per batch.
+   **/
+  const _inMemoryValues = useRef<EuiDataGridInMemoryValues>({});
+  const [inMemoryValuesVersion, setInMemoryValuesVersion] = useState(0);
 
-  const onCellRender = useCallback(
-    (rowIndex, column, value) => {
-      setInMemoryValues(inMemoryValues => {
-        const nextInMemoryValues = { ...inMemoryValues };
-        nextInMemoryValues[rowIndex] = nextInMemoryValues[rowIndex] || {};
-        nextInMemoryValues[rowIndex][column.id] = value;
-        return nextInMemoryValues;
-      });
-    },
-    [setInMemoryValues]
-  );
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  const inMemoryValues = useMemo(() => ({ ..._inMemoryValues.current }), [
+    inMemoryValuesVersion,
+  ]);
+
+  const onCellRender = useCallback((rowIndex, columnId, value) => {
+    const nextInMemoryValues = _inMemoryValues.current;
+    nextInMemoryValues[rowIndex] = nextInMemoryValues[rowIndex] || {};
+    nextInMemoryValues[rowIndex][columnId] = value;
+    setInMemoryValuesVersion(version => version + 1);
+  }, []);
 
   // if `inMemory.level` or `rowCount` changes reset the values
   const inMemoryLevel = inMemory && inMemory.level;
+  const resetRunCount = useRef(0);
   useEffect(() => {
-    setInMemoryValues({});
+    if (resetRunCount.current++ > 0) {
+      // this has to delete "overflow" keys from the object instead of resetting to an empty one,
+      // as the internal inmemoryrenderer component's useEffect which sets the values
+      // exectues before this outer, wrapping useEffect
+      const existingRowKeyCount = Object.keys(_inMemoryValues.current).length;
+      for (let i = rowCount; i < existingRowKeyCount; i++) {
+        delete _inMemoryValues.current[i];
+      }
+      setInMemoryValuesVersion(version => version + 1);
+    }
   }, [inMemoryLevel, rowCount]);
 
   return [inMemoryValues, onCellRender];