@chandlerprall 2020-06-18
> PUPPETEER_SKIP_CHROMIUM_DOWNLOAD

jenkins test this