@snide 2020-06-18
^^ ignore my comment about. The location is a React node that has links in it. The docs are clear, I just didn't notice the makeup of that column.